Added ability to us Gaussian Process Regression to data analysis. Backed up older analysis that wasn't being used.
